backend/requirements.txt
Django==4.2.11
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
pillow==12.2.0
gunicorn==23.0.0
dj-database-url==2.3.0
whitenoise==6.9.0
mysqlclient==2.2.8
